==> new_data/3-check.ipynb <==
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt

train_path = "4-train_66_withTime.csv"
test_path = "4-test_66_withTime.csv"

train = pd.read_csv(train_path)
test = pd.read_csv(test_path)
